[Device selection + VK_EXT_descriptor_indexing] Improved the device selection algorithm a bit to make it work on M1 chip.
Revisited the VK_EXT_descriptor_indexing extension
and enabled it properly. Nvidia GPU was way too
forgiving, and I was not setting up the extension
correctly. Also, on M1 it required some additional
steps as it is only in beta in MoltenVk: it is
necessary to use the
'MVK_CONFIG_USE_METAL_ARGUMENT_BUFFERS'
environment variable in order to be able to use
the extension on macOS.

I also revisited the creation of descriptor set
layouts and sets themselves because new bugs
manifested on M1. Had to redo the calculation of
descriptor counts to make it work properly.